IOT-1583: Removing /W3 warnings inside connectivity/ (part 3)
Removing the /W3 warnings in order to add the /WX option to prevent new
ones from being added inside the resource/csdk/connectivity/ directory.

Reviewed-on: https://gerrit.iotivity.org/gerrit/17549
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/csdk/connectivity/src/bt_edr_adapter/caedradapter.c
@@ -326,7 +326,7 @@ int32_t CASendEDRMulticastData(const CAEndpoint_t *endpoint, const void *data, u
     return sentLen;
 }
 
-CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, uint32_t *size)
+CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, size_t *size)
 {
     VERIFY_NON_NULL(info, TAG, "LocalConnectivity info is null");
